Improve pruning
Let's add two parameters to control the pruning behaviour:
- Excluded set: rules in this are excluded from pruning
  in order to not lose important structure.
- Use_child_rule set: by default pruned rules keep the parent
  rule. Rules in this set will be replace with the child rule
  instead.

Let's also prune nodes that have None id and no children.
This is the case for literal characters that do not carry any
information for the parent rule. For example unnamed literals
such as parenthesis can be pruned. Operators however should
not be removed as they convey meaning, but then they should
always be named (id not None).