Refactor infer_element_types
The list inference rules we want to implement are the following.
[] is empty list, [A] is list of A, [A]* is possible empty list of A

[], [] -> []
[A], [B] -> [compatible(A,B)]
[], [A] -> [A]*
[], [A]* -> [A]*
[A]*, [B]* -> [compatible(A, B)]

We can simply implement these exhaustive cases, without having to
add confusing inner recursion. These rules are still hard-coded,
but later on they can be changed to a more generic rule-based
approach that works for any set of compatible types.